Detect UTF-16 files and give a more helpful error
The TOML specification states that "A TOML file must be a valid UTF-8
encoded Unicode document", but especially on Windows it's easy to mix
this up as it calls "UTF-16" just "Unicode".

Most UTF-16 files have a BOM, so detect this and error out on it. Do
this before the lexing as lexer.next() assumes UTF-8 and mangles stuff.

Some files don't have a BOM; if a file contains NULL bytes then this is
probably the second byte in the surrogate pair.