Update to use new toml-test, fix a few bugs
Update to use the new toml-test from
https://github.com/BurntSushi/toml-test/pull/69

Turns out quite a few failures were just because those tests were
broken. Do'h.

And fix a few minor issues:

- nan and inf got encoded as "NaN" and "Inf".
- Detect invalid UTF-8 and error out.
- Don't allow keys as triple-quoted strings: """ key """ = "foo"
- Allow encoding blank keys ("" = 1) and escapes in keys ("\n" = 1).
- Expand string quoting with \b and \f, and other control characters as
  \u...
- Allow multiline strings to end with the quote character: """v""""

The toml-test tests will fail because of a catch-22: the toml-test needs
this version of the toml library to work for the fix in the lexer for
the multiline strings.